Adding base level for space Shooter
This implementation generates game window with black background. alongside generate player model and implement movement. this is done by the arrow keys and is continuous their is left, right, up, and down for the movement with boundaries. The next step is to implement the main menu and possibly connect play components to player class which is made obsolete for this implementation.